+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/extractor/impl/dml/InsertExtractor.java
@@ -15,10 +15,12 @@
  * </p>
  */
 
-package io.shardingsphere.core.parsing.antlr.extractor.impl;
+package io.shardingsphere.core.parsing.antlr.extractor.impl.dml;
 
 import com.google.common.base.Optional;
 import io.shardingsphere.core.parsing.antlr.extractor.OptionalSQLSegmentExtractor;
+import io.shardingsphere.core.parsing.antlr.extractor.impl.ColumnSegmentExtractor;
+import io.shardingsphere.core.parsing.antlr.extractor.impl.ExpressionExtractor;
 import io.shardingsphere.core.parsing.antlr.extractor.util.ExtractorUtils;
 import io.shardingsphere.core.parsing.antlr.extractor.util.RuleName;
 import io.shardingsphere.core.parsing.antlr.sql.segment.InsertSegment;
