+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/extractor/impl/dml/UpdateWhereExtractor.java
@@ -15,15 +15,18 @@
  * </p>
  */
 
-package io.shardingsphere.core.parsing.antlr.extractor.impl;
+package io.shardingsphere.core.parsing.antlr.extractor.impl.dml;
+
+import java.util.Map;
+
+import org.antlr.v4.runtime.ParserRuleContext;
 
 import com.google.common.base.Optional;
+
+import io.shardingsphere.core.parsing.antlr.extractor.impl.AbstractFromWhereExtractor;
 import io.shardingsphere.core.parsing.antlr.extractor.util.ExtractorUtils;
 import io.shardingsphere.core.parsing.antlr.extractor.util.RuleName;
 import io.shardingsphere.core.parsing.antlr.sql.segment.FromWhereSegment;
-import org.antlr.v4.runtime.ParserRuleContext;
-
-import java.util.Map;
 
 /**
  * Update where extractor.
